+++ src/Machina/Compilers/CompilerABB.cs
@@ -48,7 +48,7 @@ namespace Machina
         };
 
 
-        public override List<Types.MachinaFile> UNSAFEFullProgramFromBuffer(string programName, RobotCursor writer, bool block, bool inlineTargets, bool humanComments)
+        public override List<Types.RobotProgramFile> UNSAFEFullProgramFromBuffer(string programName, RobotCursor writer, bool block, bool inlineTargets, bool humanComments)
         {
             return null;
         }