+++ src/Machina/Compilers/CompilerMACHINA.cs
@@ -32,7 +32,7 @@ namespace Machina
 
         internal CompilerMACHINA() : base() { }
 
-        public override List<Types.MachinaFile> UNSAFEFullProgramFromBuffer(string programName, RobotCursor writer, bool block, bool inlineTargets, bool humanComments)
+        public override List<Types.RobotProgramFile> UNSAFEFullProgramFromBuffer(string programName, RobotCursor writer, bool block, bool inlineTargets, bool humanComments)
         {
             return null;
         }