+++ src/Machina/Compilers/CompilerUR.cs
@@ -38,7 +38,7 @@ namespace Machina
         internal CompilerUR() : base() { }
 
 
-        public override List<Types.MachinaFile> UNSAFEFullProgramFromBuffer(string programName, RobotCursor writer, bool block, bool inlineTargets, bool humanComments)
+        public override List<Types.RobotProgramFile> UNSAFEFullProgramFromBuffer(string programName, RobotCursor writer, bool block, bool inlineTargets, bool humanComments)
         {
             return null;
         }