+++ src/Machina/Compilers/CompilerZMORPH.cs
@@ -56,7 +56,7 @@ namespace Machina
         // Base constructor
         internal CompilerZMORPH() : base() { }
 
-        public override List<Types.MachinaFile> UNSAFEFullProgramFromBuffer(string programName, RobotCursor writer, bool block, bool inlineTargets, bool humanComments)
+        public override List<Types.RobotProgramFile> UNSAFEFullProgramFromBuffer(string programName, RobotCursor writer, bool block, bool inlineTargets, bool humanComments)
         {
             return null;
         }