+++ src/Machina/Robot.cs
@@ -11,6 +11,7 @@ using System.IO;
 using System.Runtime.Serialization.Json;
 
 using Machina.Attributes;
+using Machina.Types;
 
 //  ███╗   ███╗ █████╗  ██████╗██╗  ██╗██╗███╗   ██╗ █████╗ 
 //  ████╗ ████║██╔══██╗██╔════╝██║  ██║██║████╗  ██║██╔══██╗
@@ -309,30 +310,36 @@ namespace Machina
         }
 
         /// <summary>
-        /// Create a program in the device's native language with all the buffered Actions and return it as a nested string List,
+        /// Create a program in the device's native language with all the buffered Actions and return it as a RobotProgram,
         /// representing the different program files. Note all buffered Actions will be removed from the queue.
         /// </summary>
         /// <param name="inlineTargets">Write inline targets on action statements, or declare them as independent variables?</param>
         /// <param name="humanComments">If true, a human-readable description will be added to each line of code</param>
         /// <returns></returns>
-        public List<List<string>> Compile(bool inlineTargets = true, bool humanComments = true)
+        public RobotProgram Compile(bool inlineTargets = true, bool humanComments = true)
         {
             return c.Export(inlineTargets, humanComments);
         }
 
 
-        /// <summary>
-        /// Create a program in the device's native language with all the buffered Actions and save it to files. 
-        /// Note all buffered Actions will be removed from the queue.
-        /// </summary>
-        /// <param name="folderPath">The folder where the program files will be written to.</param>
-        /// <param name="inlineTargets">Write inline targets on action statements, or declare them as independent variables?</param>
-        /// <param name="humanComments">If true, a human-readable description will be added to each line of code</param>
-        /// <returns></returns>
-        public bool CompileToFolder(string folderPath, bool inlineTargets = true, bool humanComments = true)
+        public bool SaveProgram(RobotProgram program, string folderPath)
         {
-            return c.Export(folderPath, inlineTargets, humanComments);
+            return program.SaveToFolder(folderPath, logger);
         }
+
+
+        ///// <summary>
+        ///// Create a program in the device's native language with all the buffered Actions and save it to files. 
+        ///// Note all buffered Actions will be removed from the queue.
+        ///// </summary>
+        ///// <param name="folderPath">The folder where the program files will be written to.</param>
+        ///// <param name="inlineTargets">Write inline targets on action statements, or declare them as independent variables?</param>
+        ///// <param name="humanComments">If true, a human-readable description will be added to each line of code</param>
+        ///// <returns></returns>
+        //public bool CompileToFolder(string folderPath, bool inlineTargets = true, bool humanComments = true)
+        //{
+        //    return c.Export(folderPath, inlineTargets, humanComments);
+        //}